made the ints in Population.java long's so that it can store the big numbers

+++ src/main/java/com/napier/sem/Population.java
@@ -6,10 +6,10 @@ package com.napier.sem;
 public class Population
 {
     public String AreaName;
-    public int PopulationOfArea;
-    public int PopulationOfAreaInCities;
+    public long PopulationOfArea;
+    public long PopulationOfAreaInCities;
     public double PopulationOfAreaInCitiesPercent;
-    public int PopulationOfAreaNotInCities;
+    public long PopulationOfAreaNotInCities;
     public double PopulationOfAreaNotInCitiesPercent;
